+++ src/main/java/org/atlasapi/remotesite/pa/deletes/PaContentDeactivationPredicate.java
@@ -1,4 +1,4 @@
-package org.atlasapi.system;
+package org.atlasapi.remotesite.pa.deletes;
 
 import com.google.api.client.repackaged.com.google.common.base.Strings;
 import com.google.common.base.Function;
